pyproject.toml
[project]
name = "sql2pg_copy"
version = "2.0.0"
description = "Database migration tool with BCP, PostgreSQL COPY, flatfile, and GeoPackage support"
requires-python = ">=3.8"
dependencies = [
    "pyodbc>=4.0.39",
    "psycopg2-binary>=2.9.9",
    "PyYAML>=6.0.1",
    "colorama>=0.4.6",
    "textual>=0.47.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
gpkg = [
    "geopandas>=0.14.0",
    "fiona>=1.9.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []

